Back to only browserify
