--- controllers/controller.js
@@ -1,23 +1,0 @@
-// Require Node modules:
-var request = require('request');
-
-// Require data files:
-var storage = require('../data/storage.js');
-var sparql = require('../data/queries.js');
-var parseMultiLineString = require('../data/parseMultiLineString.js');
-
-exports.homePage = function (req, res) {
-
-}
-
-exports.searchPage = function (req, res) {
-
-}
-
-exports.detailsPage = function (req, res) {
-
-}
-
-exports.imagesPage = function (req, res) {
-
-}
