--- data/storage.js
@@ -1,10 +1,0 @@
-// Create a storage object:
-var storage = {
-	allStreets: [],
-	searchResults: [],
-	streetDetails: [],
-	streetDetailsCopy: [],
-	images: []
-};
-
-module.exports = storage;
