+++ package.json
@@ -11,7 +11,6 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.18.2",
-    "bundle-js": "^1.0.2",
     "dotenv": "^5.0.1",
     "express": "^4.16.3",
     "pug": "^2.0.3",
@@ -19,8 +18,6 @@
   },
   "devDependencies": {
     "browserify": "^16.2.2",
-    "grunt": "^1.0.2",
-    "grunt-critical": "^1.1.1",
     "nodemon": "^1.17.2"
   }
 }
